+++ scripts/test/Selenium/Agent/Admin/AdminGenericInterfaceTransportHTTPREST.t
@@ -105,6 +105,30 @@ $Selenium->RunTest(
         $Selenium->find_element( "#MaxLength", 'css' )->send_keys('1000');
         $Selenium->execute_script("\$('#KeepAlive').val('1').trigger('redraw.InputField').trigger('change');");
 
+        # Add one additional response header line.
+        $Selenium->find_element( "#AddValue", 'css' )->click();
+
+        # Click on 'Save' without entering anything to trigger client-side validation.
+        $Selenium->find_element("//button[\@value='Save and continue']")->click();
+
+        # Check if errors are shown.
+        $Self->True(
+            $Selenium->execute_script(
+                "return \$('.DefaultValueKeyItem.Error').length;"
+            ),
+            'Default key field highlighted as an error'
+        );
+        $Self->True(
+            $Selenium->execute_script(
+                "return \$('.DefaultValueItem.Error').length;"
+            ),
+            'Default key field highlighted as an error'
+        );
+
+        # Input key and value for additional response headers.
+        $Selenium->find_element( '.DefaultValueKeyItem', 'css' )->send_keys('Key1');
+        $Selenium->find_element( '.DefaultValueItem',    'css' )->send_keys('Value1');
+
         # click on 'Save'
         $Selenium->find_element("//button[\@value='Save and continue']")->VerifiedClick();
 
@@ -124,6 +148,16 @@ $Selenium->RunTest(
             $Selenium->find_element( "#KeepAlive", 'css' )->get_value(),
             1,
             "Selected value for KeepAlive field is correct"
+        );
+        $Self->Is(
+            $Selenium->find_element( '.DefaultValueKeyItem', 'css' )->get_value(),
+            'Key1',
+            'Inputed value for DefaultValueKeyItem field is correct'
+        );
+        $Self->Is(
+            $Selenium->find_element( '.DefaultValueItem', 'css' )->get_value(),
+            'Value1',
+            'Inputed value for DefaultValueItem field is correct'
         );
 
         # click on 'Save and finish' verify JS redirection