+++ var/httpd/htdocs/js/Core.Agent.Admin.GenericInterfaceTransportHTTPSOAP.js
@@ -22,6 +22,113 @@ Core.Agent.Admin = Core.Agent.Admin || {};
 Core.Agent.Admin.GenericInterfaceTransportHTTPSOAP = (function (TargetNS) {
 
     /**
+     * @name RemoveValue
+     * @memberof Core.Agent.Admin.GenericInterfaceTransportHTTPSOAP
+     * @function
+     * @returns {Boolean} Returns false.
+     * @param {String} IDSelector - ID of the pressed remove value button.
+     * @description
+     *      This function removes a value from possible values list and creates a stub input so
+     *      the server can identify if a value is empty or deleted (useful for server validation)
+     *      It also deletes the Value from the DefaultValues list
+     */
+    TargetNS.RemoveValue = function (IDSelector){
+
+        // copy HTML code for an input replacement for the deleted value
+        var $Clone = $('.DeletedValue').clone(),
+
+        // get the index of the value to delete (its always the second element (1) in this RegEx
+        $ObjectIndex = IDSelector.match(/.+_(\d+)/)[1],
+
+        // get the key name to remove it from the defaults select
+        $Key = $('#Key_' + $ObjectIndex).val();
+
+        // set the input replacement attributes to match the deleted original value
+        // new value and other controls are not needed anymore
+        $Clone.attr('id', 'Key' + '_' + $ObjectIndex);
+        $Clone.attr('name', 'Key' + '_' + $ObjectIndex);
+        $Clone.removeClass('DeletedValue');
+
+        // add the input replacement to the mapping type so it can be parsed and distinguish from
+        // empty values by the server
+        $('#' + IDSelector).closest('fieldset').append($Clone);
+
+        // remove the value from default list
+        if ($Key !== ''){
+            $('#DefaultValue').find("option[value='" + $Key + "']").remove();
+            $('#DefaultValue').trigger('redraw.InputField');
+        }
+
+        // remove possible value
+        $('#' + IDSelector).parent().remove();
+
+        return false;
+    };
+
+    /**
+     * @name AddValue
+     * @memberof Core.Agent.Admin.GenericInterfaceTransportHTTPSOAP
+     * @function
+     * @returns {Boolean} Returns false
+     * @param {Object} ValueInsert - HTML container of the value mapping row.
+     * @description
+     *      This function adds a new value to the possible values list
+     */
+    TargetNS.AddValue = function (ValueInsert) {
+
+        // clone key dialog
+        var $Clone = $('.ValueTemplate').clone(),
+            ValueCounter = $('#ValueCounter').val();
+
+        // increment key counter
+        ValueCounter++;
+
+        // remove unnecessary classes
+        $Clone.removeClass('Hidden ValueTemplate');
+
+        // add needed class
+        $Clone.addClass('ValueRow');
+
+        // copy values and change ids and names
+        $Clone.find(':input, a.RemoveButton').each(function(){
+            var ID = $(this).attr('id');
+            $(this).attr('id', ID + '_' + ValueCounter);
+            $(this).attr('name', ID + '_' + ValueCounter);
+
+            $(this).addClass('Validate_Required');
+
+            // set error controls
+            $(this).parent().find('#' + ID + 'Error').attr('id', ID + '_' + ValueCounter + 'Error');
+            $(this).parent().find('#' + ID + 'Error').attr('name', ID + '_' + ValueCounter + 'Error');
+
+            $(this).parent().find('#' + ID + 'ServerError').attr('id', ID + '_' + ValueCounter + 'ServerError');
+            $(this).parent().find('#' + ID + 'ServerError').attr('name', ID + '_' + ValueCounter + 'ServerError');
+
+            // add event handler to remove button
+            if($(this).hasClass('RemoveButton')) {
+                // bind click function to remove button
+                $(this).on('click', function () {
+                    TargetNS.RemoveValue($(this).attr('id'));
+                    return false;
+                });
+            }
+        });
+
+        $Clone.find('label').each(function(){
+            var FOR = $(this).attr('for');
+            $(this).attr('for', FOR + '_' + ValueCounter);
+        });
+
+        // prepend to container
+        ValueInsert.prepend($Clone);
+
+        // set new value for KeyName
+        $('#ValueCounter').val(ValueCounter);
+
+        return false;
+    };
+
+    /**
      * @name Init
      * @memberof Core.Agent.Admin.GenericInterfaceTransportHTTPSOAP
      * @function
@@ -101,6 +208,27 @@ Core.Agent.Admin.GenericInterfaceTransportHTTPSOAP = (function (TargetNS) {
                     $(this).removeClass('Validate_Required');
                 });
             }
+        });
+
+        // special handling for value counter
+        $('.ValueRow :input[name^="Key"]').each(function(i){
+
+            // set value counter
+            $('#ValueCounter').val(i + 1);
+        })
+
+        //bind click function to add button
+        $('#AddValue').on('click', function () {
+            TargetNS.AddValue(
+                $(this).closest('fieldset').find('.ValueInsert')
+            );
+            return false;
+        });
+
+        //bind click function to remove button
+        $('.ValueRemove').on('click', function () {
+            TargetNS.RemoveValue($(this).attr('id'));
+            return false;
         });
 
         // bind click function to Save and finish button